refactor: inline DashboardTabBar into dashboard_modal
The DashboardTabBar component is only referenced once in the
dashboard_modal layout. Inlining it simplifies the UI directory
structure and eliminates a single-use stateless widget file.
